fix: Fixed styles AuthPage and moved state of timer
